+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 2.0.0 (HEAD)
 
+* Rename "attachment" to "name", and "uploaded_file" to "attachment" in backgrounding plugin (janko-m)
+
 * Remove using `:presign` for presign options instead of `:presign_options` (janko-m)
 
 * Remove deprecated `Shrine.direct_endpoint` from direct_upload plugin (janko-m)